[트립트립] refactor(boards-list-row): 게시글 목록 훅 분리 진행
